Fixed Input form
Fixed an issue with the add a movie page where the input boxes were not populating correctly. The issue was my Javascript was dependent on a page load. Because turbo speeds up nav by performing partial page loads when possible, I changed the listener to update on "turbo:load" instead. This solved the problem.